requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
databases==0.9.0
decorator==5.1.1
distro==1.9.0
executing==2.0.1
fastapi==0.110.1
frozenlist==1.4.1
gemini-api==0.1.6
genai==2.1.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipython
jedi==0.19.1
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
openai==0.27.10
packaging==24.0
parso==0.8.4
pillow==10.3.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
regex==2023.12.25
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
yarl==1.9.4
